Transfers render films in page module
