+++ src/main.js
@@ -1,101 +1,19 @@
-import {onEscKeyDown} from './utils/common.js';
-import {renderComponent, removeComponent} from './utils/render.js';
+import {renderComponent} from './utils/render.js';
 import {Profile} from './components/profile.js';
 import {Navigation} from './components/navigation.js';
 import {Sort} from './components/sort.js';
-import {FilmCard} from './components/film-card.js';
 import {FilmList} from './components/film-list.js';
-import {ShowMoreButton} from './components/show-more-button.js';
 import {FilmStatistics} from './components/film-statistics.js';
-import {FilmInfo} from './components/film-details.js';
-import {NoData} from './components/no-data.js';
 import {generateNavigationList} from './mock/nav-list.js';
 import {generateSortList} from './mock/sort-list.js';
 import {generateFilmsCards} from './mock/film-cards.js';
 import {FILM_SECTIONS} from './mock/constants.js';
+import {PageController} from './controllers/page.js';
 
 const FILM_CARDS_AMOUNT = 20;
 const FILM_RATED_CARDS_AMOUNT = 2;
 const FILM_COMMENTED_CARDS_AMOUNT = 2;
 const BEGIN_INDEX = 0;
-const FILM_CARDS_AMOUNT_ON_START = 5;
-const FILM_CARDS_AMOUNT_LOAD_MORE = 5;
-
-const renderFilmCard = (filmsListContainer, filmCard) => {
-  const body = document.body;
-  const filmCardComponent = new FilmCard(filmCard);
-
-  const showFilmDetails = () => {
-    body.classList.add(`hide-overflow`);
-    body.appendChild(filmInfoComponent.getElement());
-    document.addEventListener(`keydown`, onDocumentEscKeyDown);
-  };
-
-  const hideFilmDetails = () => {
-    body.classList.remove(`hide-overflow`);
-    body.removeChild(filmInfoComponent.getElement());
-    document.removeEventListener(`keydown`, onDocumentEscKeyDown);
-  };
-
-  const onDocumentEscKeyDown = (evt) => {
-    onEscKeyDown(evt, hideFilmDetails);
-    document.removeEventListener(`keydown`, onDocumentEscKeyDown);
-  };
-
-  filmCardComponent.setClickHandler(showFilmDetails);
-
-  const filmInfoComponent = new FilmInfo(filmCard);
-
-  filmInfoComponent.setClickHandler(hideFilmDetails);
-
-  renderComponent(filmsListContainer, filmCardComponent);
-};
-
-const renderFilmList = (filmListComponent, filmCards) => {
-  const filmsListSection = filmListComponent.getElement().querySelector(`.films-list`);
-  const filmsListContainer = filmListComponent.getElement().querySelector(`.films-list .films-list__container`);
-  const hasFilms = filmCards.length > 0;
-
-  if (!hasFilms) {
-    renderComponent(filmsListSection, new NoData());
-  }
-
-  let showingFilmCards = FILM_CARDS_AMOUNT_ON_START;
-
-  const renderCards = (cards, container, begin, end) => {
-    cards
-    .slice(begin, end)
-    .forEach((card) => {
-      renderFilmCard(container, card);
-    });
-  };
-
-  renderCards(filmCards, filmsListContainer, BEGIN_INDEX, showingFilmCards);
-
-  const showMoreButtonComponent = new ShowMoreButton();
-
-  renderComponent(filmsListSection, showMoreButtonComponent);
-
-  showMoreButtonComponent.setClickHandler(() => {
-    const prevFilmCards = showingFilmCards;
-    showingFilmCards += FILM_CARDS_AMOUNT_LOAD_MORE;
-
-    renderCards(filmCards, filmsListContainer, prevFilmCards, showingFilmCards);
-
-    if (showingFilmCards >= filmCards.length) {
-      removeComponent(showMoreButtonComponent);
-    }
-  });
-};
-
-const renderFilmListExtra = (filmListComponent, filmCards, domElement) => {
-  const filmsListContainer = filmListComponent.getElement().querySelector(domElement);
-
-  filmCards
-    .forEach((card) => {
-      renderFilmCard(filmsListContainer, card);
-    });
-};
 
 const pageHeader = document.querySelector(`.header`);
 const pageMain = document.querySelector(`.main`);
@@ -114,8 +32,10 @@ renderComponent(pageMain, new Sort(sortList));
 
 const filmListComponent = new FilmList(FILM_SECTIONS);
 renderComponent(pageMain, filmListComponent);
-renderFilmList(filmListComponent, filmCards);
-renderFilmListExtra(filmListComponent, filmCardsTopRated, filmsTopRatedClass);
-renderFilmListExtra(filmListComponent, filmCardsMostCommented, filmsMostCommentedClass);
+
+const page = new PageController(filmListComponent);
+page.renderFilms(filmCards);
+page.renderFilmsExtra(filmCardsTopRated, filmsTopRatedClass);
+page.renderFilmsExtra(filmCardsMostCommented, filmsMostCommentedClass);
 
 renderComponent(pageFooter, new FilmStatistics());
